refactor(api): split tags.js controller into per-responsibility modules (#599)
tags.js had grown to 701 lines (audit finding C1) and mixed two concerns.
Split it by responsibility into focused routers under routes/tags/:

  crud.js     — GET/POST/PATCH/DELETE /api/tags + /api/tags/:id/{assign,unassign,assign-by-filter}
  entities.js — /api/users, /api/groups, /api/entity-tags + the *-columns-page discovery routes
  shared.js   — useSql / db + the exported ensureTagTables / buildFilterWhere / ENTITY_TO_TARGET / UUID_RE

tags.js is now a barrel that composes the two routers and re-exports the
four helpers that resources.js, admin/curatedData.js and permissions/grid.js
import from ./tags.js — so those consumers, app.js and every public path
are unchanged. Pure code move.

Grandfathered complexity entries relocated (no increase): cyclomatic
[23,22] -> crud [23] + entities [22]; cognitive [29] -> crud [29].

No new tests needed — the existing tags tests + buildFilterWhere.test.js
cover the moved code at 93%+ diff coverage, clearing the 80% gate. Full
API suite green (1752 tests, incl. resources.js's tags import); coverage
above every floor; both complexity gates and the OpenAPI-drift guard pass.
